Add files via upload
It is a program of if else statement.we use #include<iostream> it is header file. using namespace std it is our template.int main it is main function. int a it store the Integer variable. We use( a%2==0 ) it means the number which completly divisible by 2 then our if statement will execute it get even.If the number is not divisible by 2 then else statement will be executed it means odd.return 0 is for ending the program