Add files via upload
This C++ program takes two integers as input from the user, adds them together, and then prints the result. The program begins by including the iostream header file, which is required for input/output operations using cin and cout . using namespace std it is template.Three integer variables (first_number, second_number, and sum) are declared to store the input numbers and their sum.The sum of the two input numbers is calculated and stored in the variable sum.The main() function returns 0, indicating that the program ended successfully.