Add files via upload
This program  C++ is used to find the largest number from three numbers.#include<iostream> it is header file.using namespace std it is template .int main it is main function.Three integer variables a, b, and c are declared to store the numbers entered by the user.The program prompts the user to enter three numbers and stores them in variables a, b, and c using cin.The program uses three if statements to compare the numbers and determine the which is largest:The first if checks if a is greater than or equal to both b and c. If true, it prints a as the largest.The second if checks if b is greater than or equal to both a and c. If true, it prints b as the largest.The third if checks if c is greater than or equal to both a and b. If true, it prints c as the largest. Return 0  for successfuly  execution of program.