Add files via upload
This program is for the use of logical operators (&&, ||, !) in C++.#include<iostream> it is header file.using namespace std it is template.int main it is main function.Two integer variables a and b are declared and initialized with the values 8 and 5, respectively.  a && b` (Logical AND):`a` is 8 (non-zero, treated as true)
  `b` is 5 (non-zero, treated as true)
    - Result: true (1)
 `a || b` (Logical OR):
    - `a` is 8 (non-zero, treated as true)
    - `b` is 5 (non-zero, treated as true)
    - Result: true (1)
 `!b` (Logical NOT):
    - `b` is 5 (non-zero, treated as true)
    - Result: false (0), because `!` inverts the boolean value.
In C++  Non-zero values are treated as true
- Zero is treated as false
- Logical operators return 1
Return 0 is for successfuly execution of program.