Add files via upload
This program demonstrates various bitwise operations in C++. #include<ioatream> it is header file using namespace std  it is template . int main it is main function.Bitwise AND (&): Performs a logical AND between the bits of a and b, resulting in 0.  Bitwise OR (|): Performs a logical OR between the bits of a and b, resulting in 6.  Bitwise XOR (^): Performs an exclusive OR between the bits of a and b, resulting in 6.  Left Shift (<<): Shifts the bits of a one position to the left, effectively multiplying it by 2, resulting in 4. Right Shift (>>): Shifts the bits of a one position to the right, effectively dividing it by 2, resulting in 1.One's Complement (~): Flips all the bits of a, yielding -3 in two's complement form.