Add files via upload
This C++ program demonstrates the use of comparison operators ==, >, >=, <, <=, and !=m  These operators are used to compare two integers a and b #include<iostream> it is header file. Using namespace std it is template.int main it is main function. Two integer variables a and b are initialized with values 6 and 4, respectively. This compares if a is equal to b. In this case, since 6 is not equal to 4, the result is 0 (false).Then it checks if a is greater than b. Since 6 is greater than 4, the result is 1 (true). This checks if a is greater than or equal to b. Since 6 is greater than 4, the result is 1 (true). Then it checks if a is less than b. Since 6 is not less than 4, the result is 0 .This checks if a is less than or equal to b. Since 6 is not less than 4, the result is 0 (false). Then checks if a is not equal to b. Since 6 is not equal to 4, the result is 1 (true) . 
Then the program will successfully execute.